# Conflicts:
#	android/AndroidManifest.xml
#	android/src/AllFilesDocumentsProvider.java
#	android/src/MyService.java
#	android/src/NativeView.java
#	android/src/testing/MyService.java
#	android/src/testing/XCSoar.java
#	android/testing/AndroidManifest.xml
#	src/Android/NativeSensorListener.cpp

+++ android/src/SensorListener.java
@@ -58,7 +58,7 @@ public interface SensorListener {
   void onGliderLinkTraffic(long gid, String callsign,
                            double longitude, double latitude, double altitude,
                            double gspeed, double vspeed,
-                           double bearing);
+                           int bearing);
 
   void onTemperature(double temperature_kelvin);
